Added SYN program.  All builds.
SST builds using R_SYN for the SYN front end, not R_SYO.  The old R_SYO
directory has been kept for reference while modifying R_SYN to use the new
syntaxer.  R_SYN currently uses the old syntaxer, and R_SYO is not built
at all.  SST builds, and the result works as before.

Added SYN program.  This is a stripped down version of SST that only takes
SYN files as input.  This is to allow easier development of the new R_SYN
code without corrupting the SST program required for building Pascal
modules.  The intent is that building of SST be disabled in this branch
until it works again.